pyproject.toml
[tool.poetry]
name = "chatnerds"
packages = [
    {include = "chatnerds"},
]
include = ["chatnerds/config.yml", "chatnerds/config.models.yml", "chatnerds/config.prompts.yml"]
exclude = ["tests"]
version = "0.3.0"
description = "CLI application to Q&A with YouTube videos, podcasts and documents offline using local LLMs and RAG (Retrieval-augmented generation) techniques"
authors = ["Raul On Lab <raulonlab@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
langchain = ">=0.1.5,<0.2.0"
langchain-community = ">=0.0.17,<0.1.0"
llama-cpp-python = "==0.2.56"
chromadb = ">=0.4.20"
InstructorEmbedding = ">=1.0.1,<2.0.0"
sentence-transformers = "==2.2.2"  # Breaking changes in >2.2.2
huggingface_hub = ">=0.20.3"
transformers = ">=4.37.2"
auto-gptq = "==0.2.2"  # ????????????
accelerate = "^0.27.2"
openai = ">=1.11.1"
# openai-whisper = "==20231117"
openai-whisper = {git = "https://github.com/openai/whisper.git", tag = "v20231117"}
# whisper-mps = { version = "^0.0.7", python = ">=3.11" }
typing-extensions = ">=4.4.0,<5.0.0"
python-dotenv = "^0.19.2"
typer = {extras = ["all"], version = ">=0.9.0,<1.0.0"}
tqdm = ">=4.64.1,<5.0.0"
rich = "^13.0.1"
requests = "^2.31.0"
optimum = "^1.12.0"
deepmerge = ">=1.1.0,<2.0.0"
pyyaml = "^6.0"
pytubefix = "^2.3"
feedparser = "^6.0.10"
music-tag = "*"
pdfminer-six = "^20231228"
unstructured = {extras = ["all-docs"], version = "^0.12.5"}
langchain-openai = "^0.0.6"
simsimd = ">=3.7.7,<3.8"  # Breaking changes in >3.8
tiktoken = ">=0.3.3,<1"
pandoc = "^2.3"
pyfzf = "^0.3.1"
youtube-transcript-api = "^0.6.2"
qdrant-client = "^1.8.0"

[tool.poetry.group.dev.dependencies]
deptry = "*"
pylint = "*"
autoflake = "*"
black = "*"
pre-commit = "^3.6.0"
tuna = "^0.5.11"
# whisper-mps = { version = "^0.0.7", python = ">=3.11" }
pretty-errors = "^1.2.25"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
chatnerds = "chatnerds.__main__:run_cli"
